+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## next
 
+- Provide paging options for the `submission list` command.
+  - You can now specify the following options:
+  - `-l, --limit int           Limit the number of events returned (default 1)`
+  - `-o, --offset int          The number of events to offset`
+
 ## 0.0.21
 
 - Lint the Dockerfile.